refactor: 테스트 코드 추상화
